# lib/api-zod/src/generated/types/researchSession.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Pro Se Litigation Navigator API
 * OpenAPI spec version: 0.1.0
 */

export interface ResearchSession {
  id: number;
  /** @nullable */
  caseId?: number | null;
  title: string;
  messageCount: number;
  createdAt: Date;
  updatedAt: Date;
}