# lib/api-zod/src/generated/types/researchSessionWithMessages.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Pro Se Litigation Navigator API
 * OpenAPI spec version: 0.1.0
 */
import type { ResearchMessage } from './researchMessage';

export interface ResearchSessionWithMessages {
  id: number;
  /** @nullable */
  caseId?: number | null;
  title: string;
  messages: ResearchMessage[];
  createdAt: Date;
  updatedAt: Date;
}